task2 solution 2 wtf? maybe type overflow

+++ src/test/java/org/task2/SimpleEquals.java
@@ -8,64 +8,64 @@ public class SimpleEquals {
     @Test
     @DisplayName("expect 0. testcase: 1<1")
     void t1() {
-        int expected = 0;
-        int actual = Main.process("1 < 1");
+        long expected = 0;
+        long actual = Main.process("1 < 1");
         Assertions.assertEquals(expected,actual);
     }
 
     @Test
     @DisplayName("expect 0. testcase: 1>1")
     void t2() {
-        int expected = 0;
-        int actual = Main.process("1 > 1");
+        long expected = 0;
+        long actual = Main.process("1 > 1");
         Assertions.assertEquals(expected,actual);
     }
 
     @Test
     @DisplayName("expect 0. testcase: 1+1 < 1+1")
     void t3() {
-        int expected = 0;
-        int actual = Main.process("1+1 < 1+1");
+        long expected = 0;
+        long actual = Main.process("1+1 < 1+1");
         Assertions.assertEquals(expected,actual);
     }
 
     @Test
     @DisplayName("expect 0. testcase: 1+1 > 1+1")
     void t4() {
-        int expected = 0;
-        int actual = Main.process("1+1 > 1+1");
+        long expected = 0;
+        long actual = Main.process("1+1 > 1+1");
         Assertions.assertEquals(expected,actual);
     }
 
     @Test
     @DisplayName("expect 0. testcase: 2*5+5 > 5*3")
     void t5() {
-        int expected = 0;
-        int actual = Main.process("2*5+5 > 5*3");
+        long expected = 0;
+        long actual = Main.process("2*5+5 > 5*3");
         Assertions.assertEquals(expected,actual);
     }
 
     @Test
     @DisplayName("expect 0. testcase: 2*5+5 < 5*3")
     void t6() {
-        int expected = 0;
-        int actual = Main.process("2*5+5 < 5*3");
+        long expected = 0;
+        long actual = Main.process("2*5+5 < 5*3");
         Assertions.assertEquals(expected,actual);
     }
 
     @Test
     @DisplayName("expect 0. testcase: 1_000_000 < 1_000_000")
     void t7() {
-        int expected = 0;
-        int actual = Main.process("1000000 < 1000000");
+        long expected = 0;
+        long actual = Main.process("1000000 < 1000000");
         Assertions.assertEquals(expected,actual);
     }
 
     @Test
     @DisplayName("expect 1. testcase: 4*2 < 4*3 < 4*5")
     void t8() {
-        int expected = 1;
-        int actual = Main.process("4*2 < 4*3 < 4*5");
+        long expected = 1;
+        long actual = Main.process("4*2 < 4*3 < 4*5");
         Assertions.assertEquals(expected,actual);
     }
